Separate Blog Post Series component from template component
